kelas change component

+++ src/app/kelas/page.jsx
@@ -2,28 +2,30 @@
 
 import { useEffect, useState } from "react"
 import Index from "../index"
-import Table from "@/components/Table"
 import InputData from "@/components/InputData"
 import Create from "./create"
 import axios from "axios"
-import { getUrl, kelasFormat } from "../../../utils/format"
+import { deleteDialog, getNum, getUrl } from "../../../utils/format"
 import Pagination from "@/components/Pagination"
+import TableFormat, { Button, Link, Td, Tr } from "@/components/TableFormat"
+import { Delete, Edit } from "../../../public/svg"
+import PopUp from "@/components/PopUp"
 
 export default function Kelas() {
   const [dataKelas, setDataKelas] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState('');
-  const [index, setIndex] = useState(0);
   const [page, setPage] = useState(1);
+  const [view, setView] = useState(0)
   const [total, setTotal] = useState(0);
+  const [showPopUp, setPopUp] = useState(false)
 
 
   const fetchData = async () => {
     try {
       const res = await axios.get(getUrl(`/api/kelas?page=${page}`))
+      setDataKelas(res.data.kelas)
       setTotal(res.data.total);
-      const formattedData = await kelasFormat(res.data.kelas, page)
-      setDataKelas(formattedData)
     } catch (err) {
       console.error(err)
       setError(err.message)
@@ -43,27 +45,40 @@ export default function Kelas() {
       .finally(() => fetchData())
   }
 
-  const editHandler = {
-    title: "Form Edit Data Kelas",
-    indexHandler: (index) => setIndex(index),
-    form: () => <Create loading={loading} data={dataKelas[index]} submitHandler={async (data) => {
-      setLoading(true)
-      await axios.put(getUrl(`/api/kelas/${data.id}`), data)
-        .then(res => console.log(res))
-        .catch(err => console.error(err))
-        .finally(() => fetchData());
-    }} />,
-  }
+  // const editHandler = {
+  //   title: "Form Edit Data Kelas",
+  //   indexHandler: (index) => setIndex(index),
+  //   form: () => <Create loading={loading} data={dataKelas[index]} submitHandler={async (data) => {
+  //     setLoading(true)
+  //     await axios.put(getUrl(`/api/kelas/${data.id}`), data)
+  //       .then(res => console.log(res))
+  //       .catch(err => console.error(err))
+  //       .finally(() => fetchData());
+  //   }} />,
+  // }
 
-  const deleteHandler = async (id) => {
+  const editHandler = async (form) => {
     setLoading(true)
-    await axios.delete(getUrl(`/api/kelas/${id}`))
+    await axios.put(getUrl(`/api/kelas/${form.id}`), form)
       .then(res => console.log(res))
       .catch(err => {
         console.error(err)
         setError(err.response.data.message)
       })
       .finally(() => fetchData())
+  }
+
+  const deleteHandler = async (id) => {
+    if (deleteDialog()) {
+      setLoading(true)
+      await axios.delete(getUrl(`/api/kelas/${id}`))
+        .then(res => console.log(res))
+        .catch(err => {
+          console.error(err)
+          setError(err.response.data.message)
+        })
+        .finally(() => fetchData())
+    }
   }
 
   useEffect(() => {
@@ -78,7 +93,30 @@ export default function Kelas() {
           <Create loading={loading} submitHandler={submitHandler} />
         </InputData>
       </div>
-      <Table title='Kelas' data={dataKelas} loading={loading} error={error} viewHandler={'/kelas/'} editHandler={editHandler} deleteHandler={deleteHandler} />
+      <TableFormat title='Kelas' format={['No', 'Nama Kelas', 'Data Dibuat', 'Data Diubah']} loading={loading} error={error} data={dataKelas}>
+        {dataKelas.map((kelas, index) => (
+          <Tr>
+            <Td>{getNum(page, index)}</Td>
+            <Td><Link title={kelas.namaKelas} /></Td>
+            <Td><Link title={kelas.createdAt} /></Td>
+            <Td><Link title={kelas.updatedAt} /></Td>
+            <Td className='flex flex-row gap-2 justify-end'>
+              <Button clickHandler={() => {
+                setPopUp(true)
+                setView(index)
+              }} backgroundColor={'bg-orange-500'}><Edit /></Button>
+              <Button clickHandler={() => deleteHandler(siswa.id)} backgroundColor={'bg-red-500'}><Delete /></Button>
+            </Td>
+          </Tr>
+        ))
+        }
+      </TableFormat>
+      {
+        showPopUp &&
+        <PopUp title="Edit Data Kelas" onClose={() => setPopUp(false)}>
+          <Create data={dataKelas[view]} submitHandler={(form) => editHandler(form)} loading={loading} />
+        </PopUp>
+      }
       <Pagination page={page} setPage={setPage} loading={loading} total={total} />
     </Index>
   )
